Split PathHierarchy implementation into different files.
